Add coordinate uncertainty to observations

Adds a column to the observations sf object with the coordinate uncertainty in meters.

Usage

add_coordinate_uncertainty(observations, coords_uncertainty_meters = 25)

Arguments

observations

An sf object with POINT geometry.

coords_uncertainty_meters

A numeric value or a vector of numeric values indicating the coordinate uncertainty associated with observations.

Value

An sf object with POINT geometry with an additional column coordinateUncertaintyInMeters.

See also

Examples

library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

set.seed(123)

# Create four random points
n_points <- 4
xlim <- c(3841000, 3842000)
ylim <- c(3110000, 3112000)
observations_sf <- data.frame(
  lat = runif(n_points, ylim[1], ylim[2]),
    long = runif(n_points, xlim[1], xlim[2])) %>%
    st_as_sf(coords = c("long", "lat"), crs = 3035)

 # provide a fixed uncertainty for all points
 add_coordinate_uncertainty(
   observations_sf,
   coords_uncertainty_meters = 1000
   )
#> Simple feature collection with 4 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 3841046 ymin: 3110575 xmax: 3841940 ymax: 3111766
#> Projected CRS: ETRS89-extended / LAEA Europe
#>                  geometry coordinateUncertaintyInMeters
#> 1 POINT (3841940 3110575)                          1000
#> 2 POINT (3841046 3111577)                          1000
#> 3 POINT (3841528 3110818)                          1000
#> 4 POINT (3841892 3111766)                          1000

# add variability in uncertainty. For example, using gamma distribution
add_coordinate_uncertainty(
  observations_sf,
  coords_uncertainty_meters = rgamma(n_points, shape = 5, rate = 0.1)
)
#> Simple feature collection with 4 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 3841046 ymin: 3110575 xmax: 3841940 ymax: 3111766
#> Projected CRS: ETRS89-extended / LAEA Europe
#>                  geometry coordinateUncertaintyInMeters
#> 1 POINT (3841940 3110575)                      47.78440
#> 2 POINT (3841046 3111577)                      88.73564
#> 3 POINT (3841528 3110818)                      55.30862
#> 4 POINT (3841892 3111766)                      22.16495