Filter detected observations

The function filters observations from all observations based on a sampling_status column, e.g. created by sample_observations().

Usage

filter_observations(observations_total, invert = FALSE)

Arguments

observations_total

An sf object with POINT geometry or a simple dataframe with sampling_status column containing values "detected". This format is created by sample_observations().

invert

Logical. If FALSE (default), filter "detected" occurrences. Otherwise, filter all other occurrences.

Value

A dataframe or an sf object with POINT geometry containing detected occurrences (if invert = FALSE), or other occurrences (if invert = TRUE).

See also

Examples

# Load packages
library(sf)
library(dplyr)

# Set seed for reproducibility
set.seed(123)

# Simulate some occurrence data with coordinates and time points
num_points <- 10
occurrences <- data.frame(
  lon = runif(num_points, min = -180, max = 180),
  lat = runif(num_points, min = -90, max = 90),
  time_point = 0
  )

# Convert the occurrence data to an sf object
occurrences_sf <- st_as_sf(occurrences, coords = c("lon", "lat"))

# Sample observations without sampling bias
observations_total_sf <- sample_observations(
  occurrences_sf,
  detection_probability = 0.8,
  sampling_bias = "no_bias",
  seed = 123
  )

# Filter detected observations
filter_observations(observations_total_sf)
#> Simple feature collection with 7 features and 5 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -163.5997 ymin: -45.70421 xmax: 103.7898 ymax: 82.23
#> CRS:           NA
#> # A tibble: 7 × 6
#>   time_point detection_probability bias_weight sampling_probability
#>        <dbl>                 <dbl>       <dbl>                <dbl>
#> 1          0                   0.8           1                  0.8
#> 2          0                   0.8           1                  0.8
#> 3          0                   0.8           1                  0.8
#> 4          0                   0.8           1                  0.8
#> 5          0                   0.8           1                  0.8
#> 6          0                   0.8           1                  0.8
#> 7          0                   0.8           1                  0.8
#> # ℹ 2 more variables: sampling_status <chr>, geometry <POINT>

# Filter undetected observations
filter_observations(observations_total_sf, invert = TRUE)
#> Simple feature collection with 3 features and 5 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 137.8863 ymin: -82.42928 xmax: 158.5682 ymax: 13.07401
#> CRS:           NA
#> # A tibble: 3 × 6
#>   time_point detection_probability bias_weight sampling_probability
#>        <dbl>                 <dbl>       <dbl>                <dbl>
#> 1          0                   0.8           1                  0.8
#> 2          0                   0.8           1                  0.8
#> 3          0                   0.8           1                  0.8
#> # ℹ 2 more variables: sampling_status <chr>, geometry <POINT>