Simulate a random walk over time

The function simulates occurrences of a species in a temporal extent.

Usage

simulate_random_walk(
  initial_average_occurrences = 50,
  n_time_points = 10,
  sd_step = 0.05,
  seed = NA
)

Arguments

initial_average_occurrences

A positive integer value indicating the average number of occurrences to be simulated within the extend of polygon at time point 1. This value will be used as mean of a Poisson distribution (lambda parameter).

n_time_points

A positive integer value indicating the number of time points to simulate.

sd_step

A positive numeric value indicating the standard deviation of the random steps.

seed

A positive numeric value. The seed for random number generation to make results reproducible. If NA (the default), no seed is used.

Value

A vector of integers of length n_time_points with the average number of occurrences.

See also

Examples

simulate_random_walk(
  initial_average_occurrences = 50,
  n_time_points = 10,
  sd_step = 1,
  seed = 123
)
#>  [1] 50.00000 49.43952 49.20935 50.76806 50.83856 50.96785 52.68292 53.14383
#>  [9] 51.87877 51.19192